refactor: 설정 파일 및 auth 디렉토리 추가 - #19
