+++ src/main/java/saphy/saphy/cart/domain/Cart.java
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import saphy.saphy.global.entity.BaseEntity;
 import saphy.saphy.item.domain.Item;
-import saphy.saphy.user.domain.Member;
+import saphy.saphy.member.domain.Member;
 
 @Entity
 @Getter @Setter
