+++ src/main/java/saphy/saphy/member/domain/repository/UserRepository.java
@@ -1,7 +1,7 @@
-package saphy.saphy.user.domain.repository;
+package saphy.saphy.member.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import saphy.saphy.user.domain.Member;
+import saphy.saphy.member.domain.Member;
 
 public interface UserRepository extends JpaRepository<Member, Long> {
