+++ src/main/java/saphy/saphy/member/presentation/MemberController.java
@@ -1,4 +1,4 @@
-package saphy.saphy.user.presentation;
+package saphy.saphy.member.presentation;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
