+++ src/main/java/saphy/saphy/repairCenterBookmark/domain/RepairCenterBookmark.java
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import saphy.saphy.global.entity.BaseEntity;
 import saphy.saphy.repairCenter.domain.RepairCenter;
-import saphy.saphy.user.domain.Member;
+import saphy.saphy.member.domain.Member;
 
 @Entity
 @Getter
